Move blackholecluster integration test to its own folder
